requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
supabase==2.0.3
python-nmap==0.7.1
requests==2.31.0
pydantic==2.5.2
python-multipart==0.0.6
asyncio==3.4.3
aiohttp==3.9.1
cvss==2.6
